+++ src/components/Scoresheet/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Scoresheet Component Exports
+ */
+
+export { Scoresheet } from './Scoresheet'